+++ homematicip/aio/device.py
@@ -121,8 +121,14 @@ class AsyncFullFlushSwitchMeasuring(FullFlushSwitchMeasuring, AsyncSwitch):
 
 class AsyncShutterContact(ShutterContact, AsyncSabotageDevice):
     """ HMIP-SWDO (Door / Window Contact - optical) /
-    HMIP-SWDO-I (Door / Window Contact Invisible - optical) / 
+    HMIP-SWDO-I (Door / Window Contact Invisible - optical) /
     HmIP-SWDM /  HmIP-SWDM-B2  (Door / Window Contact - magnetic"""
+
+    pass
+
+
+class AsyncContactInterface(ContactInterface, AsyncSabotageDevice):
+    """ HMIP-SCI (Contact Interface Sensor) """
 
     pass