+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Devices
+  - HMIP-SCI (Contact Interface Sensor) 
 ### Fixes
 - [BUG: 223] activateVacation does not work 
+
 
 ## [0.10.9] - 2019-07-06
 ### Added